Made the compiler try to substitute indices only when they are integers or "_". The method Effects.translateMethodEffects calls Effects.substIndices and tries to substitute the formal parameters of the method by the actual arguments in those effects that have an RPL with array index RPL elements.
When DPJizer computes the types and effects symbolically, meaningless substitutions such as the ones that try to substitute an index by a noninteger variable make the substitutions unnecessarily long. Therefore, I changed the method RPL.substIndex such that it first checks if the given variable is either an integer or the default index variable "_". I applied the same change on the original DPJ compiler just to make sure that my changes are correct. I modified RPL.substIndex in the original DPJ and made sure the tests passed. I like to make the check stricter by limiting the substitution to final variables. But, when I made the compiler ignore substituting nonfinal variables, some of its unit tests failed. So, I didn't include this additional check in DPJizer.

Note that so far, I've only changed DPJizer and not DPJ because DPJ is not affected by the vacuous attempts to substitute variables.