Stopped using [?] to distinguish write effects across loop iterations.
The effects that contain the loop index variable of the foreach loop are disjoint. But, the effects that contain the RPL element [?] are not necessarily disjoint across different iterations of a foreach loop. Therefore, we changed the code so that it doesn't consider two effects containing [?] to be disjoint.

This commit belongs to issue #11.